Don't convert -ž to -h after suffix removal
Based on Firefox's machine translation and some looking up of words in
wiktionary, to me it seems the merges are almost all improvements, the
splits are mostly worse (but some are actually better), and the more
complicated cases are mostly better.

This seems particular true when the suffix removed starts with `e` -
for those there are hardly any words removing this rule harms.