C: Rework how non-localised variables are stored
Each stemmer now has its own `struct SN_local` which has any such
variables as members, with a `struct SN_env` as its first member.

We still pass around a `struct SN_env *` which now points to this first
member.  Then when we want to access a variable, we cast this pointer
to `struct SN_local *` and deference the member for the variable.

This avoid some extra allocations (one if a stemmer has any
non-localised integer or boolean variables, and another if the
stemmer has any string variables), and eliminates an indirection
on every variable access.  So it uses a bit less memory, it makes
creating and destroying a stemmer faster, and it also makes stemming
a bit faster (though only by ~0.1% for the English stemmer on our
sample vocabulary).