C++: Overhaul generated C++ code
Historically, -c++ generated exactly the same code as for C, except
with extension `.cc` instead of `.c`.  Recently it also enabled use of
`bool` for Snowball booleans.

This commit implements changes to the runtime when -c++ is used.
If malloc() fails or slick_check() fails, this is now signalled by
throwing a C++ exception, rather than returning -1 and having to
propagate that through the generated code like we do for C.  This
makes the generated code a little simpler, and more efficient in the
non-exceptional case.