Go: Fix build rule to work again
The rule is buggy because `make go` tries to make e.g.:

    go/algorithms/arabic_stemmer.go

and the rule for that actually generates:

    go/algorithms/arabic/arabic_stemmer.go

The user-visible upshot of this is that repeating `make go` will
unnecessarily regenerate all the Go code.

Unfortunately you can't use `%` multiple times in the target-pattern of
a static pattern rule - i.e. this won't work:

$(go_src_dir)/%/%_stemmer.go: $(ALGORITHMS)/%.sbl snowball$(EXEEXT)

I don't currently see a clean way to support what we want in GNU make.
Perhaps we can restructure the Go code to achieve this, but at least
for now this commit should fix CI and at least restore things to no more
broken than they have been for years (and now `make clean` actually
removes the generated Go code).